gbsim: module_id is 8 bits
A module id is an unsigned 8-bit value.  Convert all variables that
operate on module ids to use exactly that representation.

Look up the module id once in protocol handler functions, and use
that saved value.

Update format specifiers to use "%hhu" rather than "%d" when printing
module IDs.

+++ gbsim.h
@@ -132,7 +132,7 @@ static inline void gbsim_dump(__u8 *buf, size_t size)
 	fprintf(stdout, "\n");
 }
 
-static inline int cport_to_module_id(uint16_t cport)
+static inline uint8_t cport_to_module_id(uint16_t cport)
 {
 	/* FIXME can identify based on registered cport module */
 	return 1;
